refactor: restructure routes based on their usage
This is cleaner.
