+++ server/routes/version/index.ts
@@ -2,12 +2,12 @@ import fs from "fs";
 
 import express from "express";
 
-import { resolvePath } from "../lib/constants";
+import { resolvePath } from "../../lib/constants";
 
 const router = express.Router();
 
 const appInfo = JSON.parse(
-  fs.readFileSync(resolvePath(__dirname, "../package.json")).toString()
+  fs.readFileSync(resolvePath(__dirname, "../../package.json")).toString()
 );
 
 router.get("/", (_req, res) =>
